requirements.txt
setuptools>=69.0.0
highrise-bot-sdk>=24.1.0
aiohttp>=3.9.0
yt-dlp[default]>=2024.10.0
imageio-ffmpeg>=0.5.1
